refactor(cli): wrap tablegen logic into a standalone function
